+++ packages/ui/src/parts/time-zones/routes/time-zones.$.tsx
@@ -1,10 +1,10 @@
 import { createFileRoute, notFound, redirect } from '@tanstack/react-router';
-import { getLocale } from '@joeblew999/remy-ui/locale';
-import { m } from '@joeblew999/remy-ui/messages';
-import { pageHead } from '@joeblew999/remy-ui/tanstack';
-import { TimeZonePage, canonicalTimeZone, timeZoneName, timeZonePath } from '@joeblew999/remy-ui/showcase/time-zone';
-import { ErrorPage, Problem } from '../problem';
-import { usePreferred } from '../preferred';
+import { getLocale } from '../../../locale';
+import { m } from '../../../paraglide/messages.js';
+import { pageHead } from '../../../tanstack';
+import { TimeZonePage, canonicalTimeZone, timeZoneName, timeZonePath } from '../../../showcase/time-zone';
+import { ErrorPage, Problem } from '../../../problem';
+import { usePreferred } from '../../../preferred';
 
 // A sub-resource of the formats page: /es/time-zones/Asia/Tokyo. The splat holds the IANA name;
 // a name Intl does not know throws notFound() (a localized 404 naming it), and another spelling